Carry a caller's row identifier through the in-memory indexes as an int64
rtree_insert, sptree_insert, their temporal and split variants and the two
nearest-neighbour cursors take the identifier as an int, while a caller
identifies a row with an int64: DuckDB's row_t is int64, so a binding narrows
every identifier it stores and the index answers with a truncated one.

Widen the identifier to int64 through the whole path: the six insert entry
points, rtree_nn_cursor_next and sptree_nn_cursor_next, the ids array of an
RTreeNode and the id of an SPTree node, and every MeosArray that a search, a
join or a dedup fills. meos_array_add copies elem_size bytes, so an array
created at one width and read at another is silent rather than a compile error;
the element size and the cast move together at each of the sites that consume a
search result.

Collapse the duplicates of a split search by sorting rather than by a set
indexed with the identifier. A set sized to the largest identifier allocates in
proportion to a value the caller chooses, which reaches gigabytes on a sparse
identifier space and has no bound at 64 bits; sorting costs memory proportional
to the number of candidates.

+++ meos/examples/rtree_example.c
@@ -105,7 +105,7 @@ verify_search(const char *name, const RTree *rtree, RTreeSearchOp op,
   bool *indexed = calloc(NUM_BBOX, sizeof(bool));
   for (int i = 0; i < index_count; i++)
   {
-    int id = *(int *) meos_array_get(ids, i);
+    int64 id = *(int64 *) meos_array_get(ids, i);
     indexed[id] = true;
   }
 
@@ -391,7 +391,7 @@ int main(void)
   printf("RTree index test (%d boxes per type)\n", NUM_BBOX);
 
   /* Create a single MeosArray and reuse it across all searches */
-  MeosArray *ids = meos_array_create(sizeof(int));
+  MeosArray *ids = meos_array_create(sizeof(int64));
   test_floatspan(ids);
   test_tstzspan(ids);
   test_tbox(ids);
